backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
python-docx
google-genai
python-dotenv
pydantic
